8: progress

+++ 7/1/utils.h
@@ -4,4 +4,4 @@ extern void cloneArray(const long size, long from[], long to[]);
 
 extern long areSame(const long size, const long arr1[], const long arr2[]);
 
-extern long arrayIsOrdered(const long size, const long arr[]);
+extern int arrayIsOrdered(const long size, const long arr[]);
